LayoutTests/fast/dom/SelectorAPI/closest-definition.html

<!doctype html>
<html>
<head>
<script src="../../../resources/js-test-pre.js"></script>
</head>
<body>
</body>
<script>
description('Test the definition of Element.closest().');

shouldBeUndefined('Element.closest');
shouldBeTrue('Object.getOwnPropertyDescriptor(Element.prototype, "closest").enumerable');
shouldBeTrue('Object.getOwnPropertyDescriptor(Element.prototype, "closest").writable');
shouldBe('Element.prototype.closest.length', '1');

</script>
<script src="../../../resources/js-test-post.js"></script>
</html>
