Safari-611.1.10/LayoutTests/fast/dom/SelectorAPI/closest-definition-expected.txt

Test the definition of Element.closest().

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Element.closest is undefined.
PASS Object.getOwnPropertyDescriptor(Element.prototype, "closest").enumerable is true
PASS Object.getOwnPropertyDescriptor(Element.prototype, "closest").writable is true
PASS Element.prototype.closest.length is 1
PASS successfullyParsed is true

TEST COMPLETE
